src/main/java/com/example/tdd/controller/ClientController.java
package com.example.tdd.controller;

import com.example.tdd.entity.ClientEntity;
import com.example.tdd.services.ClientService;
import com.example.tdd.services.IClientService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.context.annotation.Bean;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.Optional;

@Slf4j
@RestController
@RequestMapping("/api/clients")
public class ClientController {

    @Autowired
    private IClientService clientService;

    @GetMapping(path = "/", produces = "application/json")
    public ResponseEntity<Object> findAll() throws Exception {
        List<ClientEntity> clients = clientService.findAll();
        if (clients.size() == 0)
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        return new ResponseEntity<>(clients, HttpStatus.OK);
    }

    @GetMapping("/{id}")
    public ResponseEntity<Object> findById(@PathVariable Long id)  {
        Optional<ClientEntity> client= clientService.findById(id);
        if(client.isEmpty())
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        return new ResponseEntity<>(client, HttpStatus.OK);
    }


}
